requirements.txt
diffusers>=0.3.0
SwissArmyTransformer==0.4.11 # Inference
torch==2.4.0
torchvision==0.19.0
gradio==4.40.0 # For HF gradio demo
streamlit==1.37.0 # For web demo
opencv-python==4.10.0.82
imageio-ffmpeg==0.5.1
openai==1.38.0
transformers==4.43.4
accelerate==0.33.0
sentencepiece==0.2.0
pillow==9.5.0